+++ lib/kweerie/result_class_generator.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
 require_relative "result_class_components/accessors"
-require_relative "result_class_components/attributes"
-require_relative "result_class_components/changes"
 require_relative "result_class_components/comparison"
 require_relative "result_class_components/key_transformation"
 require_relative "result_class_components/serialization"
@@ -13,14 +11,29 @@ module Kweerie
     def self.generate(parent_class, attribute_names)
       Class.new(parent_class) do
         include Comparable
-        include ResultClassComponents::Attributes
-        include ResultClassComponents::Serialization
-        include ResultClassComponents::Comparison
-        include ResultClassComponents::Changes
         include ResultClassComponents::Accessors
+        include ResultClassComponents::Comparison
+        include ResultClassComponents::KeyTransformation
+        include ResultClassComponents::Serialization
+        include ResultClassComponents::TypeCasting
 
         # Define attr_readers for all columns
         attribute_names.each { |name| attr_reader name }
+
+        define_method :initialize do |attrs|
+          # Store both raw and casted versions
+          cast_definitions = parent_class.cast_definitions
+
+          @_raw_original_attributes = attrs.dup
+          @_original_attributes = attrs.each_with_object({}) do |(key, value), hash|
+            type_definition = cast_definitions[key.to_sym]
+            casted_value = type_cast_value(value, type_definition)
+            hash[key.to_s] = casted_value
+            instance_variable_set("@#{key}", casted_value)
+          end
+
+          super() if defined?(super)
+        end
 
         # Nice inspect output
         define_method :inspect do