+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## [Unreleased]
+## [0.2.0] - 2024-11-12
+
+- changed BaseObjects to BaseObject.
+- remove automatic type inference.
+- added cast_select.
+- added tiny amound of custom types.
 
 ## [0.1.0] - 2024-11-06